Convert SalvageGood to ManufacturingManager
Changes:
- SalvageGood Task moved to correct package
- SalvageGoods now accepted the target Building in the constructor
- Remove unused methds from SalvageGood
- ManufacturingMetaTask create backlog taks that support the SalvageGood Task that takes a Building and optional Salvagable object
- Remove the queuing of Salvage processes fromt eh Manufacture function; queuing takes place in the central manager

Part of #1236
